Restructuring facts for clearer testing
I want to start with the input data and check some facts about it.
Perform some operations on the input data and check some facts about
that.
This is very much a threading kind of thing.  So, I made a function that
takes two functions as arguments.  One function to check some facts and
a second to move along the thread.